Summarization will now be done by a celery worker
